requirements.txt
django
djangorestframework
pygments
requests>=2.13.0
selenium>=3.4.0
asyncio==3.4.3